Break state out into own file
