+++ Node/.gitignore
@@ -13,5 +13,7 @@ Core/Src
 !Core/Inc/can_handler.h
 !Core/Inc/utilities.h
 !Core/Inc/can_message_defs.h
+!Core/Src/state.c
+!Core/Inc/state.h
 compile_commands.json
 .cache/
